+++ fotoobo/tools/fc/get.py
@@ -3,7 +3,7 @@ FortiCloud get utility
 """
 
 import logging
-from typing import Any
+from typing import Any, Dict, List
 
 from fotoobo.fortinet.forticloud import FortiCloudAsset
 from fotoobo.helpers.config import config
@@ -13,17 +13,26 @@ from fotoobo.inventory import Inventory
 log = logging.getLogger("fotoobo")
 
 
-def products(host: str) -> Result[Any]:
+def products(host: str) -> Result[List[Dict[str, Any]]]:
     """
-    Return the assets from Asset management in FortiCloud"""
-    result = Result[Any]()
+    FortiCloud API get products from Asset management
+
+    Here we assume that just one page is returned. If it happens that more that one page is returned
+    pagination has to be implemented.
+
+    Args:
+        host: Host defined in inventory
+
+    Returns:
+        The asset list as a result object
+    """
+    result = Result[List[Dict[str, Any]]]()
     inventory = Inventory(config.inventory_file)
     fc: FortiCloudAsset = inventory.get_item(host, "forticloud")
     log.debug("FortiCloud get assets ...")
     response = fc.post("/products/list", payload={"serialNumber": "%"})
-
-    if "error" in response and response["error"]:
-        result.push_message(host, response["error"]["message"], "error")
+    if response["error"]:
+        result.push_message(host, response["message"], "info")
 
     else:
         result.push_result(host, response["assets"])
@@ -39,7 +48,7 @@ def version(host: str) -> Result[str]:
         host: Host defined in inventory
 
     Returns:
-        The version string for the FortiCloud API
+        The version string for the FortiCloud API in a result object
     """
     result = Result[str]()
     inventory = Inventory(config.inventory_file)